+++ src/wreck/lock_loader.py
@@ -27,20 +27,14 @@ Both implementations hold the result in memory.
 multiple implementations based on the first arg, not the return type
 
 .. py:data:: __all__
-   :type: tuple[str, str, str, str]
-   :value: ("LoaderImplementation", "LoaderPin", "LoaderPinDatum", \
-   "from_loader_filepins")
+   :type: tuple[str, str, str]
+   :value: ("LoaderImplementation", "LoaderPinDatum", "from_loader_filepins")
 
    Module exports
 
 """
 
 import abc
-
-from pip_requirements_parser import (
-    InstallationError,
-    RequirementsFile,
-)
 
 from .constants import (
     SUFFIX_IN,
@@ -49,7 +43,6 @@ from .constants import (
 from .exceptions import MissingRequirementsFoldersFiles
 from .lock_datum import (
     DATUM,
-    Pin,
     PinDatum,
     is_pin,
 )
@@ -62,7 +55,6 @@ from .pep518_venvs import (
 
 __all__ = (
     "LoaderImplementation",
-    "LoaderPin",
     "LoaderPinDatum",
     "from_loader_filepins",
 )
@@ -89,104 +81,6 @@ def _check_filter_by_pin(val, default=True):
         ret = val
 
     return ret
-
-
-def _from_loader_pins(
-    loader,
-    venv_path,
-    suffix=SUFFIX_UNLOCKED,
-    filter_by_pin=True,
-):
-    """Factory. From a venv, get all Pins.
-
-    :param loader: Contains some paths and loaded not parsed venv reqs
-    :type loader: wreck.pep518_venvs.VenvMapLoader
-    :param venv_path:
-
-       Relative path to venv base folder. Acts as a key
-
-    :type venv_path: typing.Any
-    :param suffix:
-
-       Default ``.unlock``. End suffix of compiled requirements file.
-       Either ``.unlock`` or ``.lock``
-
-    :type suffix: str
-    :param filter_by_pin: Default True. Filter out entries without specifiers
-    :type filter_by_pin: bool | None
-    :returns: Feed list[Pin] into class constructor to get an Iterator[Pin]
-    :rtype: set[wreck.lock_inspect._T]
-    :raises:
-
-        - :py:exc:`wreck.exceptions.MissingRequirementsFoldersFiles` --
-          missing requirements file(s). Create it
-
-        - :py:exc:`wreck.exceptions.MissingPackageBaseFolder` --
-          loader not provided. Loader provides package base folder
-
-
-    .. deprecated:: 0.0.1
-
-       drain_swamp.lock_inspect is still around as a reference (bad) implementation
-       There is no wreck.lock_inspect
-
-    """
-    check_loader(loader)
-    filter_by_pin = _check_filter_by_pin(filter_by_pin)
-
-    """Relative path acts as a dict key. An absolute path is not a key.
-    Selecting by always nonexistent key returns empty Sequence, abspath_reqs"""
-    relpath_venv = check_venv_relpath(loader, venv_path)
-
-    # NotADirectoryError, ValueError, KeyError, MissingRequirementsFoldersFiles
-    abspath_reqs = get_reqs(loader, relpath_venv, suffix_last=suffix)
-
-    pins = set()
-    for abspath_req in abspath_reqs:
-
-        """Take only enough details to identify package as a pin.
-
-        In .unlock files, ``-c`` and ``-r`` are resolved. Therefore
-        ``options`` section will be empty
-        """
-        try:
-            rf = RequirementsFile.from_file(abspath_req)
-        except InstallationError as exc:
-            msg_exc = (
-                f"For venv {venv_path!s}, requirements file not "
-                f"found {abspath_req!r}. Create it"
-            )
-            raise MissingRequirementsFoldersFiles(msg_exc) from exc
-
-        d_rf_all = rf.to_dict()
-        rf_reqs = d_rf_all["requirements"]
-        has_reqs = (
-            rf_reqs is not None and isinstance(rf_reqs, list) and len(rf_reqs) != 0
-        )
-        if has_reqs:
-            for d_req in rf_reqs:
-                pkg_name = d_req.get("name", "")
-                # list[specifier]
-                specifiers = d_req.get("specifier", [])
-
-                # Only keep package which have specifiers e.g. ``pip>=24.2``
-                if filter_by_pin is True:
-                    if is_pin(specifiers):
-                        pin = Pin(abspath_req, pkg_name)
-                        pins.add(pin)
-                    else:  # pragma: no cover
-                        # non-Pin --> filtered out
-                        pass
-                else:
-                    # Do not apply filter. All entries
-                    pin = Pin(abspath_req, pkg_name)
-                    pins.add(pin)
-        else:  # pragma: no cover
-            pass
-    else:  # pragma: no cover
-        pass
-
-    return pins
 
 
 def from_loader_filepins(
@@ -351,41 +245,6 @@ class LoaderImplementation(abc.ABC):
         :rtype: set[wreck.lock_datum.DATUM]
         """
         ...
-
-
-class LoaderPin(LoaderImplementation):
-    """Original implementation
-
-    Inefficient cuz reads the file, then reads the file for every pin
-
-    Usage
-
-    .. code-block:: text
-
-       ret: set[wreck.lock_datum.DATUM] = LoaderPin()(loader, venv_path)
-
-    .. deprecated:: 0.0.1
-
-       drain_swamp.lock_inspect is still around as a reference (bad) implementation
-       There is no wreck.lock_inspect
-
-    """
-
-    def __call__(
-        self,
-        loader,
-        venv_path,
-        suffix=SUFFIX_UNLOCKED,
-        filter_by_pin=True,
-    ) -> set[DATUM]:
-        """Returns a Pin. For signature See the abc"""
-        set_ret = _from_loader_pins(
-            loader,
-            venv_path,
-            suffix=suffix,
-            filter_by_pin=filter_by_pin,
-        )
-        return set_ret
 
 
 class LoaderPinDatum(LoaderImplementation):